requirements.txt
# === Core AI ===
openai>=1.35.0

# === Web API ===
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9  # For file uploads

# === File Parsing (Simplified - No OCR) ===
pymupdf>=1.24.2  # PDF extraction
python-docx>=1.1.0  # Word document (.docx) support

# === Environment & Utils ===
python-dotenv>=1.0.1
requests>=2.31.0

# === Optional: for better performance ===
watchfiles>=0.21.0
httpx>=0.27.0

# ============================================
# REMOVED (No longer needed):
# ============================================
# pillow - Was for image processing
# pytesseract - Was for OCR (image text extraction)
# langchain - Was for LangGraph autograder
# langgraph - Was for complex graph workflows
# sympy - Was for math evaluation
